refactor: inline platform helpers
Consolidate android and iOS helper logic into the primary FlutterInappPurchase entry point and remove the legacy modules directory. Update examples, tests, and documentation to reflect the new generated handler usage while adding convenience VS Code launchers.
