+++ example/lib/test_storekit.dart
@@ -56,10 +56,17 @@ class _TestScreenState extends State<TestScreen> {
 
       // Test 3: Get simple product
       setState(() => _status = 'Getting products...');
-      final products = await _iap.fetchProducts(
-        skus: ['dev.hyo.martie.10bulbs'],
-        type: ProductType.InApp,
+      final result = await _iap.fetchProducts(
+        ProductRequest(
+          skus: const ['dev.hyo.martie.10bulbs'],
+          type: ProductQueryType.InApp,
+        ),
       );
+
+      final products = result is FetchProductsResultProducts
+          ? result.value ?? const <Product>[]
+          : const <Product>[];
+
       setState(() => _status = 'Got ${products.length} products');
 
       if (products.isNotEmpty) {
