+++ lib/utils.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+
 import 'package:flutter/foundation.dart';
 
 /// Parse Android JSON response and log the result
