+++ test/fetch_products_all_test.dart
@@ -15,36 +15,46 @@ void main() {
         case 'initConnection':
           return true;
         case 'fetchProducts':
-          return <Map<String, dynamic>>[
-            <String, dynamic>{
-              'platform': 'ios',
-              'id': 'premium_monthly',
-              'type': 'subs',
-              'title': 'Premium Monthly',
-              'description': 'Monthly subscription',
-              'currency': 'USD',
-              'displayNameIOS': 'Premium Monthly',
-              'displayPrice': '\$24.99',
-              'isFamilyShareableIOS': false,
-              'jsonRepresentationIOS': '{}',
-              'typeIOS': 'AUTO_RENEWABLE_SUBSCRIPTION',
-              'price': 24.99,
-            },
-            <String, dynamic>{
-              'platform': 'ios',
-              'id': 'coin_pack',
-              'type': 'in-app',
-              'title': 'Coin Pack',
-              'description': 'One time coins',
-              'currency': 'USD',
-              'displayNameIOS': 'Coin Pack',
-              'displayPrice': '\$2.99',
-              'isFamilyShareableIOS': false,
-              'jsonRepresentationIOS': '{}',
-              'typeIOS': 'NON_CONSUMABLE',
-              'price': 2.99,
-            },
-          ];
+          final args = call.arguments as Map<dynamic, dynamic>?;
+          final type = args?['type']?.toString();
+
+          final subscription = <String, dynamic>{
+            'platform': 'ios',
+            'id': 'premium_monthly',
+            'type': 'subs',
+            'title': 'Premium Monthly',
+            'description': 'Monthly subscription',
+            'currency': 'USD',
+            'displayNameIOS': 'Premium Monthly',
+            'displayPrice': '\$24.99',
+            'isFamilyShareableIOS': false,
+            'jsonRepresentationIOS': '{}',
+            'typeIOS': 'AUTO_RENEWABLE_SUBSCRIPTION',
+            'price': 24.99,
+          };
+
+          final inApp = <String, dynamic>{
+            'platform': 'ios',
+            'id': 'coin_pack',
+            'type': 'in-app',
+            'title': 'Coin Pack',
+            'description': 'One time coins',
+            'currency': 'USD',
+            'displayNameIOS': 'Coin Pack',
+            'displayPrice': '\$2.99',
+            'isFamilyShareableIOS': false,
+            'jsonRepresentationIOS': '{}',
+            'typeIOS': 'NON_CONSUMABLE',
+            'price': 2.99,
+          };
+
+          if (type == 'subs') {
+            return <Map<String, dynamic>>[subscription];
+          }
+          if (type == 'inapp') {
+            return <Map<String, dynamic>>[inApp];
+          }
+          return <Map<String, dynamic>>[subscription, inApp];
       }
       return null;
     });
@@ -55,24 +65,66 @@ void main() {
         .setMockMethodCallHandler(channel, null);
   });
 
-  test('fetchProducts detects per-item type when querying all products',
+  test('fetchProducts returns in-app products when querying in-app type',
       () async {
     final platform = FakePlatform(operatingSystem: 'ios');
     final iap = FlutterInappPurchase.private(platform);
 
     await iap.initConnection();
 
-    final result = await iap.fetchProducts<types.ProductCommon>(
-      skus: const ['premium_monthly', 'coin_pack'],
-      type: types.ProductQueryType.All,
+    final result = await iap.fetchProducts(
+      const types.ProductRequest(
+        skus: ['premium_monthly', 'coin_pack'],
+        type: types.ProductQueryType.InApp,
+      ),
     );
 
-    final subs = result.whereType<types.ProductSubscriptionIOS>().toList();
-    final inApps = result.whereType<types.ProductIOS>().toList();
+    expect(result, isA<types.FetchProductsResultProducts>());
+    final products =
+        (result as types.FetchProductsResultProducts).value ?? const [];
+    expect(products, hasLength(1));
+    expect(products.first.id, 'coin_pack');
+  });
 
-    expect(subs, hasLength(1));
-    expect(subs.first.id, 'premium_monthly');
-    expect(inApps, hasLength(1));
-    expect(inApps.first.id, 'coin_pack');
+  test('fetchProducts returns subscriptions when querying subs type', () async {
+    final platform = FakePlatform(operatingSystem: 'ios');
+    final iap = FlutterInappPurchase.private(platform);
+
+    await iap.initConnection();
+
+    final result = await iap.fetchProducts(
+      const types.ProductRequest(
+        skus: ['premium_monthly', 'coin_pack'],
+        type: types.ProductQueryType.Subs,
+      ),
+    );
+
+    expect(result, isA<types.FetchProductsResultSubscriptions>());
+    final subscriptions =
+        (result as types.FetchProductsResultSubscriptions).value ?? const [];
+    expect(subscriptions, hasLength(1));
+    expect(subscriptions.first.id, 'premium_monthly');
+  });
+
+  test('fetchProducts throws when querying with ProductQueryType.All',
+      () async {
+    final platform = FakePlatform(operatingSystem: 'ios');
+    final iap = FlutterInappPurchase.private(platform);
+
+    await iap.initConnection();
+
+    await expectLater(
+      () => iap.fetchProducts(
+        const types.ProductRequest(
+          skus: ['premium_monthly', 'coin_pack'],
+          type: types.ProductQueryType.All,
+        ),
+      ),
+      throwsA(isA<types.PurchaseError>().having(
+        (error) => error.code,
+        'code',
+        types.ErrorCode.DeveloperError,
+      )),
+    );
   });
 }
